refactor: split dom.js into focused modules to reduce coupling
- Extracted setupKeyboardShortcuts → src/utils/keyboard-helpers.js
- Extracted setupInlineInput, startInlineRename → src/utils/form-helpers.js
- Extracted setupDropZone → src/utils/drop-zone-helpers.js
- Kept core in dom.js: _el, createActionButton, renderButtonBar, buildChevronRow, createModalOverlay
- Added backward-compat re-exports in dom.js (temporary)
- Updated all importing files to use the new module paths
