+++ src/components/webview-panel.js
@@ -1,4 +1,5 @@
-import { _el, setupKeyboardShortcuts } from '../utils/dom.js';
+import { _el } from '../utils/dom.js';
+import { setupKeyboardShortcuts } from '../utils/keyboard-helpers.js';
 import { trackMouse } from '../utils/drag-helpers.js';
 import {
   MAX_LOGS,
