+++ src/utils/context-menu.js
@@ -1,5 +1,6 @@
-import { _el, positionInViewport, setupKeyboardShortcuts } from './dom.js';
+import { _el, positionInViewport } from './dom.js';
 import { onClickStopped } from './event-helpers.js';
+import { setupKeyboardShortcuts } from './keyboard-helpers.js';
 
 export class ContextMenu {
   constructor() {
